fix(pg): parallel build infra — PGDLLEXPORT symbol, size_t math, per-participant DSM share
Three independent fixes uncovered while running SIFT1M with
parallel_workers > 0:

1. graph_index_parallel_build_main was extern "C" but not PGDLLEXPORT.
   With -fvisibility=hidden the symbol stays internal in pg_vexdb.so;
   PG's LookupParallelWorkerFunction dlsym fails -> "could not find
   function". Add PGDLLEXPORT on both decl and def.

2. bulkbuf_smgr.cpp calculate_pow: `targe_size_mb * 1024 * 1024` did
   the math in uint32. With maintenance_work_mem=16GB (targe_size_mb
   = 16184) the product wraps to ~3.8GB before the divide, producing
   wrong chunk sizes. Use size_t * 1024ULL * 1024ULL.

3. graph_index_build.cpp pre-allocation: both leader and each worker
   computed `mempool_initsize_mb = total - 200` and emplaced a MemStore
   that bump_alloc'd one chunk per pool against the shared DSM
   segment. Five participants x three pools => up to 30 GB requested
   from a 16 GB DSM -> "Allocator out of memory" on the first worker
   init. Divide the initial-pool target by `nparticipants` on both
   sides; add `nparticipants` to GraphIndexShared so workers see the
   leader's count.

Also (separate, in include/pq.h): PQ scan's get_distance_single was
indexing dist_table with raw float bytes interpreted as PQ codes
(storage holds raw vectors, not codes -- on-disk encode is the
follow-up). That OOB index triggered SIGSEGV the moment any query
hit the PQ index. Route get_distance_single and get_distance_batch2
through `_get_distance_precise_func` until PQ codes really live on
disk; behavior matches plain HNSW, no crash.

Verified on the SIFT1M / PG 19 test rig:
- Single-thread CREATE INDEX (no parallel_workers reloption) builds
  cleanly, ~30 min for 1M x 128d, recall@10 = 1.000 at ef_search=400.
- Parallel build (parallel_workers=4) starts cleanly with all 5
  participants pre-allocating and gets past the OOM that previously
  killed it within seconds; but during the actual graph build a
  separate SIGSEGV appears -- the MemPool design has per-participant
  `Vector<Chunk> vec` (not in DSM), so worker A's inserted nodes are
  not addressable from worker B by NodeId, and B's HNSW search reads
  uninitialised memory. That is an architectural bug (shared
  store/atomic id allocator) tracked as the next follow-up; this
  commit only fixes the surface bugs above.

PQ scan SIGSEGV is gone.

+++ src/bulkbuf_smgr.cpp
@@ -13,7 +13,8 @@ using namespace ann_helper;
 uint32 calculate_pow(uint32 elem_size, uint32 targe_size_mb)
 {
     targe_size_mb = Max(targe_size_mb, 1);
-    uint32 target_size = targe_size_mb * 1024 * 1024ul;
+    // 64-bit math: targe_size_mb >= 4096 would overflow uint32 here.
+    size_t target_size = (size_t)targe_size_mb * 1024ULL * 1024ULL;
     double target_count = static_cast<double>(target_size) / elem_size;
     double x_double = log2(target_count);
     int x = static_cast<int>(floor(x_double));
